Refactor the whole script in order for the module to be loaded more efficiently
